[#10364] move HasUuid to dedicated package
