+++ sormas-api/src/main/java/de/symeda/sormas/api/EntityDto.java
@@ -23,6 +23,7 @@ import java.util.Date;
 import javax.validation.constraints.Pattern;
 import javax.validation.constraints.Size;
 
+import de.symeda.sormas.api.uuid.HasUuid;
 import org.apache.commons.lang3.StringUtils;
 
 import com.fasterxml.jackson.annotation.JsonIgnore;
