+++ sormas-api/src/main/java/de/symeda/sormas/api/feature/FeatureConfigurationIndexDto.java
@@ -1,6 +1,6 @@
 package de.symeda.sormas.api.feature;
 
-import static de.symeda.sormas.api.HasUuid.UUID_REGEX;
+import static de.symeda.sormas.api.uuid.HasUuid.UUID_REGEX;
 import static de.symeda.sormas.api.utils.FieldConstraints.CHARACTER_LIMIT_UUID_MAX;
 import static de.symeda.sormas.api.utils.FieldConstraints.CHARACTER_LIMIT_UUID_MIN;
