+++ sormas-api/src/main/java/de/symeda/sormas/api/sormastosormas/share/incoming/RequestResponseDataDto.java
@@ -23,7 +23,7 @@ import java.io.Serializable;
 import javax.validation.constraints.Pattern;
 import javax.validation.constraints.Size;
 
-import de.symeda.sormas.api.HasUuid;
+import de.symeda.sormas.api.uuid.HasUuid;
 import de.symeda.sormas.api.i18n.Validations;
 import de.symeda.sormas.api.utils.FieldConstraints;
