+++ sormas-api/src/main/java/de/symeda/sormas/api/sormastosormas/share/incoming/SormasToSormasEventParticipantPreview.java
@@ -22,7 +22,7 @@ import java.io.Serializable;
 
 import javax.validation.Valid;
 
-import de.symeda.sormas.api.HasUuid;
+import de.symeda.sormas.api.uuid.HasUuid;
 import de.symeda.sormas.api.event.EventReferenceDto;
 import de.symeda.sormas.api.utils.EmbeddedPersonalData;
 import de.symeda.sormas.api.utils.EmbeddedSensitiveData;
