+++ sormas-api/src/main/java/de/symeda/sormas/api/uuid/HasUuid.java
@@ -15,7 +15,7 @@
  * You should have received a copy of the GNU General Public License
  * along with this program. If not, see <https://www.gnu.org/licenses/>.
 */
-package de.symeda.sormas.api;
+package de.symeda.sormas.api.uuid;
 
 public interface HasUuid {
