+++ sormas-backend/src/main/java/de/symeda/auditlog/api/Auditor.java
@@ -40,7 +40,7 @@ import de.symeda.auditlog.api.value.format.ValueFormatter;
 import de.symeda.auditlog.api.value.format.override.DateFormatOverrideDetector;
 import de.symeda.auditlog.api.value.format.override.OverrideDetector;
 import de.symeda.auditlog.api.value.reflection.EntityInspector;
-import de.symeda.sormas.api.HasUuid;
+import de.symeda.sormas.api.uuid.HasUuid;
 
 /**
  * Class for the inspection of entity states.
