+++ sormas-backend/src/main/java/de/symeda/auditlog/api/value/format/DefaultCollectionFormatter.java
@@ -20,7 +20,7 @@ package de.symeda.auditlog.api.value.format;
 import java.io.Serializable;
 import java.util.Comparator;
 
-import de.symeda.sormas.api.HasUuid;
+import de.symeda.sormas.api.uuid.HasUuid;
 
 /**
  * Supports the following elements in Collections (relevant for a reliable sorting):
