+++ sormas-backend/src/main/java/de/symeda/sormas/backend/common/CriteriaBuilderHelper.java
@@ -19,7 +19,7 @@ import javax.persistence.criteria.Predicate;
 import org.apache.commons.collections.CollectionUtils;
 import org.apache.commons.collections4.ListUtils;
 
-import de.symeda.sormas.api.HasUuid;
+import de.symeda.sormas.api.uuid.HasUuid;
 import de.symeda.sormas.api.utils.DataHelper;
 import de.symeda.sormas.backend.ExtendedPostgreSQL94Dialect;
 import de.symeda.sormas.backend.util.ModelConstants;
