+++ sormas-backend/src/test/java/de/symeda/auditlog/api/sample/CollectionEntity.java
@@ -25,7 +25,7 @@ import java.util.Set;
 
 import de.symeda.auditlog.api.Audited;
 import de.symeda.auditlog.api.AuditedIgnore;
-import de.symeda.sormas.api.HasUuid;
+import de.symeda.sormas.api.uuid.HasUuid;
 
 @Audited
 public class CollectionEntity implements HasUuid {
