+++ sormas-backend/src/test/java/de/symeda/auditlog/api/sample/Entity.java
@@ -19,7 +19,7 @@ package de.symeda.auditlog.api.sample;
 
 import de.symeda.auditlog.api.Audited;
 import de.symeda.auditlog.api.AuditedIgnore;
-import de.symeda.sormas.api.HasUuid;
+import de.symeda.sormas.api.uuid.HasUuid;
 
 @Audited
 public class Entity implements HasUuid {
