+++ sormas-backend/src/test/java/de/symeda/auditlog/api/sample/EntityWithIgnoredMethods.java
@@ -26,7 +26,7 @@ import javax.persistence.Transient;
 import de.symeda.auditlog.api.Audited;
 import de.symeda.auditlog.api.AuditedAttribute;
 import de.symeda.auditlog.api.AuditedIgnore;
-import de.symeda.sormas.api.HasUuid;
+import de.symeda.sormas.api.uuid.HasUuid;
 
 @Audited
 public class EntityWithIgnoredMethods implements HasUuid {
