+++ sormas-backend/src/test/java/de/symeda/auditlog/api/sample/OverridingFormatterEntity.java
@@ -24,7 +24,7 @@ import javax.persistence.TemporalType;
 
 import de.symeda.auditlog.api.Audited;
 import de.symeda.auditlog.api.AuditedIgnore;
-import de.symeda.sormas.api.HasUuid;
+import de.symeda.sormas.api.uuid.HasUuid;
 
 @Audited
 public class OverridingFormatterEntity implements HasUuid {
